simple-contact-form/requirements.txt
asgiref==3.7.2
coverage==7.2.7
crispy-bootstrap4==2022.1
Django==4.2.2
django-crispy-forms==2.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
inflection==0.3.1
iniconfig==2.0.0
lazy-object-proxy==1.9.0
model-bakery==1.12.0
packaging==23.1
pluggy==1.0.0
psycopg2-binary==2.9.6
pytest==7.3.2
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-drf==1.1.3
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
wrapt==1.15.0
